scraper/requirements.txt
playwright==1.58.0
supabase==2.29.0
python-dotenv==1.2.2
